refactor: minimize file footprints of FiniteNumber; fix test
