--- src/domain/primitives/FiniteNumber/FiniteNumber.type.ts
@@ -1,1 +1,0 @@
-export type FiniteNumber = number & { readonly _tag: unique symbol };
