+++ src/domain/primitives/FiniteNumber/FiniteNumber.types.ts
@@ -1,0 +1,9 @@
+import { DomainError } from '@domain/types';
+
+export type FiniteNumberError = DomainError & {
+  code: 'FiniteNumberError';
+  reason: 'OutOfRange';
+  message: 'FiniteNumber must be a finite number';
+};
+
+export type FiniteNumber = number & { readonly _tag: unique symbol };
