+++ src/domain/primitives/index.ts
@@ -1,9 +1,9 @@
-export type { FiniteNonZeroPositiveNumber } from './FiniteNonZeroPositiveNumber/FiniteNonZeroPositiveNumber.type.ts';
-export type { FiniteNumber } from './FiniteNumber/FiniteNumber.type.ts';
 export type {
   NonEmptyString,
   NonEmptyStringError,
 } from './NonEmptyString/NonEmptyString.types.ts';
+export type { FiniteNonZeroPositiveNumber } from './FiniteNonZeroPositiveNumber/FiniteNonZeroPositiveNumber.type.ts';
+export type { FiniteNumber, FiniteNumberError } from './FiniteNumber/FiniteNumber.types.ts';
 
 export { createFiniteNonZeroPositiveNumber } from './FiniteNonZeroPositiveNumber/FiniteNonZeroPositiveNumber.factory.ts';
 export { createFiniteNumber } from './FiniteNumber/FiniteNumber.factory.ts';
